+++ frontend/components/base/base-range/BaseRangeSlider.vue
@@ -76,6 +76,8 @@ export default {
 };
 </script>
 <style lang="scss" scoped>
+$slider-color: $primary-color;
+$slider-light-color: rgba(0, 26, 255, 0.1);
 label {
   width: fit-content;
   height: 14px;
@@ -97,13 +99,26 @@ label {
     @include font-size(12px);
   }
   &__progress-value {
+    display: none;
     position: absolute;
     top: $base-space * 3;
-    margin-left: -20px;
-    width: 40px;
+    margin-left: -15px;
+    width: 30px;
     text-align: center;
-    color: $black-54;
+    color: palette(white);
+    background: palette(grey, 200);
+    border-radius: $border-radius-s;
     @include font-size(12px);
+    &:before {
+      position: absolute;
+      left: calc(50% - 6px);
+      top: 0;
+      transform: translateY(-50%);
+      @include triangle(top, 6px, 6px, palette(grey, 200));
+    }
+    .range:hover & {
+      display: block;
+    }
   }
   &__slider {
     -webkit-appearance: none;
@@ -122,7 +137,7 @@ input[type="range"]::-webkit-slider-thumb {
   appearance: none;
   height: 15px;
   width: 15px;
-  background-color: $primary-color;
+  background-color: $slider-color;
   border-radius: 50%;
   border: none;
   transition: 0.2s ease-in-out;
@@ -131,28 +146,28 @@ input[type="range"]::-webkit-slider-thumb {
 input[type="range"]::-moz-range-thumb {
   height: 15px;
   width: 15px;
-  background-color: $primary-color;
+  background-color: $slider-color;
   border-radius: 50%;
   border: none;
   transition: 0.2s ease-in-out;
 }
 
 input[type="range"]::-webkit-slider-thumb:hover {
-  box-shadow: 0 0 0 6px rgba(0, 26, 255, 0.1);
+  box-shadow: 0 0 0 4px $slider-light-color;
 }
 input[type="range"]:active::-webkit-slider-thumb {
-  box-shadow: 0 0 0 10px rgba(0, 26, 255, 0.1);
+  box-shadow: 0 0 0 6px $slider-light-color;
 }
 input[type="range"]:focus::-webkit-slider-thumb {
-  box-shadow: 0 0 0 10px rgba(0, 26, 255, 0.1);
+  box-shadow: 0 0 0 6px $slider-light-color;
 }
 input[type="range"]::-moz-range-thumb:hover {
-  box-shadow: 0 0 0 6px rgba(0, 26, 255, 0.1);
+  box-shadow: 0 0 0 4px $slider-light-color;
 }
 input[type="range"]:active::-moz-range-thumb {
-  box-shadow: 0 0 0 10px rgba(0, 26, 255, 0.1);
+  box-shadow: 0 0 0 6px $slider-light-color;
 }
 input[type="range"]:focus::-moz-range-thumb {
-  box-shadow: 0 0 0 10px rgba(0, 26, 255, 0.1);
+  box-shadow: 0 0 0 6px $slider-light-color;
 }
 </style>