+++ frontend/components/feedback-task/settings/SettingsQuestions.vue
@@ -48,6 +48,28 @@
               "
             >
               <label :for="`options-${question.id}`" v-text="$t('labels')" />
+              <BaseSwitch
+                v-if="question.isMultiLabelType"
+                class="settings__edition-form__switch --subcategory"
+                :id="`options-order-${question.id}`"
+                v-model="question.settings.suggestionFirst"
+                >{{ $t("suggestionFirst") }}</BaseSwitch
+              >
+
+              <BaseRangeSlider
+                v-if="question.settings.shouldShowVisibleOptions"
+                class="settings__edition-form__slider --subcategory"
+                :id="`visible_options-${question.id}`"
+                :min="3"
+                :max="question.settings.options.length"
+                v-model="question.settings.visible_options"
+                >{{ $t("visibleLabels") }}</BaseRangeSlider
+              >
+
+              <label
+                v-text="$t('order')"
+                class="settings__edition-form__label --subcategory"
+              />
               <draggable
                 class="label__container"
                 ghost-class="label__item__ghost"
@@ -73,18 +95,10 @@
 
             <BaseSwitch
               v-if="question.isTextType"
+              class="settings__edition-form__switch"
               :id="`use-markdown-${question.id}`"
               v-model="question.settings.use_markdown"
               >{{ $t("useMarkdown") }}</BaseSwitch
-            >
-
-            <BaseRangeSlider
-              v-if="question.settings.options?.length > 3"
-              :id="`visible_options-${question.id}`"
-              :min="3"
-              :max="question.settings.options.length"
-              v-model="question.settings.visible_options"
-              >{{ $t("visibleLabels") }}</BaseRangeSlider
             >
 
             <div class="settings__edition-form__footer">
@@ -156,11 +170,20 @@ export default {
     overflow: auto;
     padding-right: $base-space * 2;
 
+    &__label {
+      &.--subcategory {
+        display: block;
+        @include font-size(13px);
+        color: $black-87;
+      }
+    }
+
     &__content {
       display: flex;
       flex-direction: column;
       gap: $base-space * 2;
       width: 100%;
+      max-width: 800px;
       margin-top: $base-space;
     }
 
@@ -184,16 +207,50 @@ export default {
       }
     }
 
+    &__switch.re-switch {
+      :deep(label) {
+        color: $black-87;
+      }
+      @include media(">desktop") {
+        :deep(label) {
+          min-width: 140px;
+        }
+      }
+      &.--subcategory {
+        margin-top: $base-space * 2;
+        @include font-size(13px);
+      }
+    }
+
+    &__slider {
+      :deep(label) {
+        color: $black-87;
+      }
+      @include media(">desktop") {
+        :deep(label) {
+          min-width: 140px;
+          margin-right: 1.5em;
+        }
+      }
+      &.--subcategory {
+        @include font-size(13px);
+        @include media(">desktop") {
+          display: flex;
+          align-items: center;
+        }
+      }
+    }
+
     &__group {
       display: flex;
       flex-direction: column;
       width: 100%;
-      gap: $base-space;
+      gap: 12px;
 
       & > label {
         width: fit-content;
         height: 14px;
-        color: $black-54;
+        color: $black-87;
       }
 
       & input {