+++ frontend/v1/infrastructure/services/useRunningEnvironment.ts
@@ -1,10 +1,6 @@
 import { useResolve } from "ts-injecty";
+import { GetEnvironmentUseCase } from "~/v1/domain/usecases/get-environment-use-case";
 import { OAuthLoginUseCase } from "~/v1/domain/usecases/oauth-login-use-case";
-
-interface HFSpace {
-  space: string;
-  user: string;
-}
 
 const HUGGING_FACE_EMBEBED_URL = "huggingface.co";
 const HUGGING_FACE_DIRECT_URL = ".hf.space";
@@ -12,41 +8,35 @@ const HUGGING_FACE_DIRECT_URL = ".hf.space";
 export const useRunningEnvironment = () => {
   const url = new URL(window.location.href);
 
-  const parseHuggingFaceData = (data: string[]) => {
-    const user = data[0];
-    const space = data[1];
-
-    if (data.length === 2 && user && space)
-      return {
-        space,
-        user,
-      };
-
-    return undefined;
-  };
-
   const isEmbebed = () => {
     return window.self !== window.top;
   };
 
-  const isRunningOnHuggingFace = (): HFSpace | undefined => {
-    if (url.host === HUGGING_FACE_EMBEBED_URL) {
-      const paramsData = url.pathname
-        .replace("/spaces", "")
-        .split("/")
-        .filter(Boolean);
+  const isRunningOnHuggingFace = (): boolean => {
+    return (
+      url.host === HUGGING_FACE_EMBEBED_URL ||
+      url.host.endsWith(HUGGING_FACE_DIRECT_URL)
+    );
+  };
 
-      return parseHuggingFaceData(paramsData);
-    }
+  const getEnvironment = async () => {
+    const environmentUseCase = useResolve(GetEnvironmentUseCase);
 
-    if (url.host.endsWith(HUGGING_FACE_DIRECT_URL)) {
-      const paramsData = url.host
-        .replaceAll(HUGGING_FACE_DIRECT_URL, "")
-        .split(/-(.*)/s)
-        .filter(Boolean);
+    return await environmentUseCase.execute();
+  };
 
-      return parseHuggingFaceData(paramsData);
-    }
+  const hasPersistentStorageWarning = async () => {
+    if (!isRunningOnHuggingFace()) return false;
+
+    const environment = await getEnvironment();
+
+    return environment.shouldShowHuggingfaceSpacePersistantStorageWarning;
+  };
+
+  const getHuggingFaceSpace = async () => {
+    const environment = await getEnvironment();
+
+    return environment.huggingFaceSpace;
   };
 
   const hasHuggingFaceOAuthConfigured = async (): Promise<boolean> => {
@@ -60,6 +50,9 @@ export const useRunningEnvironment = () => {
   return {
     isEmbebed,
     isRunningOnHuggingFace,
+    getEnvironment,
+    getHuggingFaceSpace,
+    hasPersistentStorageWarning,
     hasHuggingFaceOAuthConfigured,
   };
 };